common/utils/test-mqtt/requirements.txt
paho-mqtt
kafka-python
docker